Add files via upload
This is a sample for a volatility indicator (future correlation is 10%) and it is just for showcasing ability to implement research and tangible results to that implementation. I do not take any responsibility what so ever for what you use it for. The volatility estimate was done via EWMA (GARCH(1,1))